server/requirements/base.txt
Click==7.0
coreapi==2.3.3
Django==2.1.1
django-cors-headers==2.4.0
djangorestframework==3.8.2
Pillow==5.2.0
